add clean cmd and refactor build:index
