+++ packages/dendron-plugin-views/scripts/start.js
@@ -38,31 +38,6 @@ const react = require(require.resolve('react', { paths: [paths.appPath] }));
 const env = getClientEnvironment(paths.publicUrlOrPath.slice(0, -1));
 const useYarn = fs.existsSync(paths.yarnLockFile);
 const isInteractive = process.stdout.isTTY;
-const {WebViewCommonUtils} = require("@dendronhq/common-server")
-
-const _ = require('lodash');
-const path = require("path");
-
-// Compile Dendron `index.html` template
-let theme = process.env.THEME || "";
-
-const out = WebViewCommonUtils.genVSCodeHTMLIndex({
-  // dummy, not used. for browser mode, this is added by CRA app
-  jsSrc: "",
-  cssSrc: "",
-  // cssSrc: `${path.join("public", "static", "css", theme + ".css")}`,
-  port: 3005,
-  wsRoot: path.resolve(path.join("..", "..", "test-workspace")),
-  browser: true,
-  acquireVsCodeApi: `window.vscode = {postMessage: ()=>{}};`,
-  themeMap: {
-    light: `${path.join("static", "css", "themes", "light.css")}`,
-    dark: `${path.join("static", "css", "themes", "dark.css")}`,
-  },
-  initialTheme: theme,
-});
-fs.writeFileSync(path.join("public/index.html"), out);
-
 
 // Warn and crash if required files are missing
 if (!checkRequiredFiles([paths.appHtml, paths.appIndexJs])) {
